+++ assets/javascripts/kitten/components/cards/crowdfunding-card/components/state.js
@@ -2,10 +2,10 @@ import React, { PureComponent } from 'react'
 import styled from 'styled-components'
 import PropTypes from 'prop-types'
 import COLORS from '../../../../constants/colors-config'
-import { Text } from '../../../../components/typography/text'
 import { pxToRem } from '../../../../helpers/utils/typography'
 import { KissKissBankBankIcon } from '../../../../components/icons/kisskissbankbank-icon'
 import { ScreenConfig } from '../../../../constants/screen-config'
+import { Text } from '../../../../components/typography/text'
 
 const COMPONENT_GUTTER = pxToRem(10)
 
@@ -35,19 +35,10 @@ const StyledWidgetContainer = styled.div`
   padding: ${pxToRem(10)} ${pxToRem(20)};
 `
 
-const StyledWidgetText = styled(Text)`
-  display: flex;
-  align-items: center;
-`
-
-const StyledWidgetLogo = styled(KissKissBankBankIcon)`
-  margin-left: ${pxToRem(5)};
-`
-
 class State extends PureComponent {
   static propTypes = {
     state: PropTypes.string,
-    widgetState: PropTypes.string,
+    widgetState: PropTypes.node,
     loading: PropTypes.bool,
   }
 
@@ -95,17 +86,7 @@ class State extends PureComponent {
     return (
       <>
         {!loading && (
-          <StyledWidgetContainer>
-            <StyledWidgetText
-              size="micro"
-              lineHeight="normal"
-              weight="regular"
-              color="background1"
-            >
-              {widgetState}
-              <StyledWidgetLogo width="80" />
-            </StyledWidgetText>
-          </StyledWidgetContainer>
+          <StyledWidgetContainer>{widgetState}</StyledWidgetContainer>
         )}
       </>
     )