+++ assets/javascripts/kitten/components/cards/crowdfunding-card/widget/index.js
@@ -10,12 +10,15 @@ import Loading from '../components/loading'
 import State from '../components/state'
 import { pxToRem } from '../../../../helpers/utils/typography'
 
+const COMPONENT_GUTTER = pxToRem(10)
+
 const StyledContainer = styled.div`
   position: relative;
 `
 
 const StyledTitleAndDescription = styled.div`
-  margin-top: ${pxToRem(10)};
+  padding: 0 ${COMPONENT_GUTTER};
+  margin-top: ${pxToRem(5)};
 
   ${({ titlesMinHeight }) =>
     titlesMinHeight &&