+++ assets/javascripts/kitten/karl/widget/stories.js
@@ -1,0 +1,7 @@
+import React from 'react'
+import { storiesOf } from '@storybook/react'
+import WidgetCard from './components/widget-card'
+
+storiesOf('Widget/Widget-card', module).add('Widget Card', () => {
+  return <WidgetCard />
+})